backend/requirements.txt
Flask>=2.3.0,<3.0.0
Flask-SQLAlchemy>=3.0.0,<4.0.0
Flask-CORS>=4.0.0,<5.0.0
Flask-Compress>=1.14,<2.0.0
Flask-Limiter>=3.5.0,<4.0.0
requests>=2.31.0,<3.0.0